implement remaining nodes

+++ extension/src/nodes/decorators/beehave_cooldown.cpp
@@ -47,7 +47,7 @@ void BeehaveCooldown::_bind_methods() {
 	ClassDB::bind_method(D_METHOD("get_wait_time"), &BeehaveCooldown::get_wait_time);
 
 	// exports
-	ADD_PROPERTY(PropertyInfo(Variant::INT, "wait_time"), "set_wait_time", "get_wait_time");
+	ADD_PROPERTY(PropertyInfo(Variant::FLOAT, "wait_time"), "set_wait_time", "get_wait_time");
 }
 
 void BeehaveCooldown::set_wait_time(float wait_time) {
